un peu de random dans ce monde de brute

+++ Sources/Ennemi.cpp
@@ -40,7 +40,7 @@ bool Ennemi::isDead()
 	return lifePoint <= 0;
 }
 
-int Ennemi::takeHit(int strenght)
+float Ennemi::takeHit(float strenght)
 {
 	return 0;
 }
